+++ client/src/pages/SignUpPage/index.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import jwt_decode from 'jwt-decode';
 import { useSetRecoilState } from 'recoil';
 import { toast } from 'react-toastify';
-import { Button, Modal } from '@/lib/design';
-import rightArrow from '@public/icons/arrow-right.svg';
+
+import { Modal } from '@/lib/design';
 import * as S from './style';
 import userAtom from '@/recoil/user';
 import { errorMessage } from '@/lib/common/message';
@@ -11,10 +11,8 @@ import { NewUser, signUp } from '@/lib/api/user';
 import { UserState } from '@/contexts/userContext';
 import { taskSortByUpdate } from '@/lib/utils/sort';
 import { searchOrganizationByName } from '@/lib/api/organization';
-import * as avatar from '@/lib/common/avatar';
-import { ImageType } from '@/types/image';
-
-const MAXINDEX = 16;
+import AvatarForm from '@/components/AvatarForm';
+import SignUpForm from '@/components/SignUpForm';
 
 const SignUpPage = ({ token }: { token: string }) => {
   const { room, email }: { room: string; email: string } =
@@ -34,13 +32,6 @@ const SignUpPage = ({ token }: { token: string }) => {
   const [avatarIndex, setAvatarIndex] = useState(0);
   const [showModal, setShowModal] = useState(false);
   const [context, setContext] = useState('');
-  const changeGender = () => (gender === 'Boy' ? setGender('Girl') : setGender('Boy'));
-
-  const clickLeft = () =>
-    !avatarIndex ? setAvatarIndex(MAXINDEX) : setAvatarIndex((prev) => prev - 1);
-
-  const clickRight = () =>
-    avatarIndex === MAXINDEX ? setAvatarIndex(0) : setAvatarIndex((prev) => prev + 1);
 
   const createUserByInput = async () => {
     if (newUser.password !== newUser.checkPassword) {
@@ -85,58 +76,18 @@ const SignUpPage = ({ token }: { token: string }) => {
         <S.Text> {context}</S.Text>
       </Modal>
       <S.Title to="/">HyupUp</S.Title>
-      <S.AvatarSelectContainer>
-        <button onClick={clickLeft}>
-          <S.LeftArrow src={rightArrow} color="red" />
-        </button>
-        <S.AvatarContainer>
-          <S.Avatar src={avatar[`${gender}${avatarIndex}` as unknown as ImageType]}></S.Avatar>
-          <Button type="button" category="default" size="small" onClick={changeGender}>
-            {gender}
-          </Button>
-        </S.AvatarContainer>
-        <button onClick={clickRight}>
-          <S.RightArrow src={rightArrow} alt="" />
-        </button>
-      </S.AvatarSelectContainer>
-      <S.FormBox onSubmit={checkOrganizationByName}>
-        <S.InputBox
-          placeholder="이름"
-          value={newUser.name}
-          onChange={(e) => setNewUser(() => ({ ...newUser, name: e.target.value }))}
-        ></S.InputBox>
-        <S.InputBox
-          placeholder="직무"
-          value={newUser.job}
-          onChange={(e) => setNewUser(() => ({ ...newUser, job: e.target.value }))}
-        ></S.InputBox>
-        <S.InputBox
-          placeholder="이메일"
-          type="email"
-          value={newUser.email}
-          onChange={(e) => setNewUser(() => ({ ...newUser, email: e.target.value }))}
-        ></S.InputBox>
-        <S.InputBox
-          placeholder="비밀번호"
-          type="password"
-          value={newUser.password}
-          onChange={(e) => setNewUser(() => ({ ...newUser, password: e.target.value }))}
-        ></S.InputBox>
-        <S.InputBox
-          placeholder="비밀번호를 다시 입력해주세요"
-          type="password"
-          value={newUser.checkPassword}
-          onChange={(e) => setNewUser(() => ({ ...newUser, checkPassword: e.target.value }))}
-        ></S.InputBox>
-        <S.InputBox
-          placeholder="조직 이름"
-          value={newUser.organization}
-          onChange={(e) => setNewUser(() => ({ ...newUser, organization: e.target.value }))}
-        ></S.InputBox>
-        <Button category="default" size="large">
-          HyupUp 시작하기
-        </Button>
-      </S.FormBox>
+      <AvatarForm
+        gender={gender}
+        avatarIndex={avatarIndex}
+        setGender={setGender}
+        setAvatarIndex={setAvatarIndex}
+      />
+      <SignUpForm
+        token={token}
+        checkOrganizationByName={checkOrganizationByName}
+        newUser={newUser}
+        setNewUser={setNewUser}
+      />
     </S.Container>
   );
 };
